Refactor model structure
