+++ lib/load_games.dart
@@ -1,9 +1,12 @@
 import 'dart:convert';
 import 'package:http/http.dart' as http;
-import 'package:how_many_mobile_meeple/settings.dart';
+import 'package:how_many_mobile_meeple/model/settings.dart';
 
 import 'game_config.dart';
-import 'model.dart';
+import 'package:how_many_mobile_meeple/model/item.dart';
+
+import 'model/game.dart';
+import 'model/games.dart';
 
 class LoadGames {
   static Future<Games> fetchGames(Settings settings, List<Item> items) async {
@@ -23,68 +26,5 @@ class LoadGames {
     }
 
     return games;
-  }
-}
-
-class Games {
-  final Map<String, Game> gamesByName;
-
-  List<Game> get games => gamesByName.values.toList();
-
-  Games({this.gamesByName});
-
-  factory Games.fromJson(List<dynamic> parsedJson) {
-    var games = Map<String, Game>();
-
-    games = Map.fromEntries(parsedJson.map((gameData) {
-      var gameFromJs = Game.fromJson(gameData);
-      return MapEntry(gameFromJs.name, gameFromJs);
-    }));
-
-    return new Games(
-      gamesByName: games,
-    );
-  }
-
-  Games addGames(Games newGames) {
-    gamesByName.addAll(newGames.gamesByName);
-    return this;
-  }
-
-  List<Game> getGamesByRating() {
-    List<Game> unsortedGames = games;
-    unsortedGames.sort((a, b) => b.averageRating.compareTo(a.averageRating));
-    return unsortedGames;
-  }
-}
-
-class Game {
-  final String name;
-  final int maxPlayers;
-  final int minPlayers;
-  final int maxPlaytime;
-  final String imageUrl;
-  final String thumbnailUrl;
-  final double averageRating;
-
-  Game(
-      {this.name,
-      this.maxPlayers,
-      this.minPlayers,
-      this.maxPlaytime,
-      this.imageUrl,
-      this.thumbnailUrl,
-      this.averageRating});
-
-  factory Game.fromJson(Map<String, dynamic> json) {
-    return Game(
-      name: json['name'],
-      maxPlayers: json['maxplayers'],
-      minPlayers: json['minplayers'],
-      maxPlaytime: json['maxplaytime'],
-      imageUrl: json['image'],
-      thumbnailUrl: json['thumbnail'],
-      averageRating: json['stats']['average'] ?? 0,
-    );
   }
 }
