+++ lib/main.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 
 import 'package:scoped_model/scoped_model.dart';
 
-import 'package:how_many_mobile_meeple/model.dart';
+import 'package:how_many_mobile_meeple/model/model.dart';
 import 'package:how_many_mobile_meeple/random_game_display.dart';
 
 import 'homepage.dart';
