+++ lib/model/settings.dart
