+++ lib/random_game_display.dart
@@ -6,10 +6,11 @@ import 'package:how_many_mobile_meeple/screen_tools.dart';
 
 import 'app_default_padding.dart';
 import 'app_page.dart';
+import 'package:how_many_mobile_meeple/model/bgg_cache.dart';
 import 'game_config.dart';
 import 'how_many_meeple_app_bar.dart';
-import 'load_games.dart';
-import 'model.dart';
+import 'package:how_many_mobile_meeple/model/model.dart';
+import 'model/game.dart';
 import 'network_content_widget.dart';
 
 class RandomGameDisplayPage extends NetworkWidget with AppPage {
